backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
anthropic==0.28.1
pyyaml==6.0.1
